init Clock component: jsx, scss and tests
The Clock.jsx component which renders the time is created. With prop
totalSeconds, the component process the raw seconds into minute:second
format. When the totalseconds is passed in as props, the components
calls this.formatSeconds and render the result(xx:xx) on the page.

The tests are created to check the process of passed totalseconds value
into the minute:second format is working as intended.

With the scss, the component is styled.