update countdown: init countdownForm
The form takes seconds and when it is submitted, it calls for its
onSubmit method. This validates the input data if it is a number, then
pass it as a integer to props.onSetCountdown which is connected to
parent component(Countdown.jsx)'s handleSetCountdown. This set state
'count' as passed in seconds, and then pass the data to Clock.jsx. So
the input number by user is passed to Clock.jsx.