update tests: testing if the method is working
With spy function, it checks when the form is submitted the intended
results whether happens.

The spy checks the activity flow. This test simulates when the form is
activated with the value 109 is input. And as intended, the submitted
number is passed in to props, which is onSetCountdown={spy} in CountdownForm.