updated Countdown component: init timer function
In Countdown.jsx, with componentDiDUpdate method when countdownStatus
status is changed the startTimer method is executed. This method runs
setInterval which takes a function and interval as parameter. The first
argument extracts 1 from current count state and update it through
setState(). This happens in every 1 second and passed into Clock
component each time when the state is updated.