init Controls.jsx and its test:
The Controls component is for controling the timer. When the timer is
started, this control component appears and enables to pause or clear
the timer.

This component is handled by the props called countdownStatus which
takes state of countdownStatus of Countdown component. When status is
set to 'started', it shows pause button and when status is 'paused', it
shows start button.

The condition is written as function inside the render method and called
inside the html.

In the current test, it tests for an existence of the component.