update Controls & Countdown: linking and actioning
Mouting(rendering) buttons depends on the state(countdownsStatus). With
each button's onClick function it updates status of parent component(in
this case Countdown.jsx), and by taking advantage of react lifecycle
method 'componentDidUpdate' which is activated when state changes,
functions are added inside.

When state is set to stopped, it reset the count to zero and run paused
case which clears inverval(which means stopping countdown). When it is paused it only
stops countdown. Note that the `break` is intentionally taken out.