update Countdown: clearInterval
With componentWillUnmount lifecycle method, the function that counts
down number stops right before the component is unmounted. In this case
clearInterval is used to clear the function out.

Also in startTimer, to prevent the timer function to be executed even
after the count reaches 0, added conditional which changes
countdownStatus to 'stopped', so once it reaches 0, it doesn't longer
function.